Shuffling composition/balance types to feature_table
