--- q2_types/composition/__init__.py
@@ -1,11 +1,0 @@
-# ----------------------------------------------------------------------------
-# Copyright (c) 2016-2017, QIIME 2 development team.
-#
-# Distributed under the terms of the Modified BSD License.
-#
-# The full license is in the file LICENSE, distributed with this software.
-# ----------------------------------------------------------------------------
-from ._type import Composition, Balance
-
-
-__all__ = ["Composition", "Balance"]
