--- q2_types/composition/_type.py
@@ -1,20 +1,0 @@
-# ----------------------------------------------------------------------------
-# Copyright (c) 2016-2017, QIIME 2 development team.
-#
-# Distributed under the terms of the Modified BSD License.
-#
-# The full license is in the file LICENSE, distributed with this software.
-# ----------------------------------------------------------------------------
-from qiime2.plugin import SemanticType
-from q2_types.feature_table import FeatureTable
-from ..plugin_setup import plugin
-
-
-Composition = SemanticType('Composition',
-                           variant_of=FeatureTable.field['content'])
-
-Balance = SemanticType('Balance',
-                       variant_of=FeatureTable.field['content'])
-
-
-plugin.register_semantic_types(Composition, Balance)
