+++ q2_types/feature_table/__init__.py
@@ -10,10 +10,11 @@ import importlib
 
 from ._format import (BIOMV100Format, BIOMV210Format, BIOMV100DirFmt,
                       BIOMV210DirFmt)
-from ._type import FeatureTable, Frequency, RelativeFrequency, PresenceAbsence
+from ._type import (FeatureTable, Frequency, RelativeFrequency, PresenceAbsence,
+                    Composition, Balance)
 
 __all__ = ['BIOMV100Format', 'BIOMV100DirFmt', 'FeatureTable', 'Frequency',
            'RelativeFrequency', 'PresenceAbsence', 'BIOMV210Format',
-           'BIOMV210DirFmt']
+           'BIOMV210DirFmt', 'Composition', 'Balance']
 
 importlib.import_module('q2_types.feature_table._transformer')
